requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-cloud-vision==0.42.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
opencv-python==4.1.2.30
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
